requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.0.2
openai==1.3.7
PyPDF2==3.0.1
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
tenacity==8.2.3
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
